feat(mobile): add display preference override to account address
