+++ apps/mobile/src/features/accounts/components/account-address.tsx
@@ -1,12 +1,21 @@
-import { useAccountDisplayAddress } from '@/store/settings/settings.read';
+import { useAccountDisplayAddress } from '@/hooks/use-account-display-address';
 
-import { AccountId } from '@leather.io/models';
+import { AccountDisplayPreference } from '@leather.io/models';
 import { Text, TextProps } from '@leather.io/ui/native';
 
-type AccountAddressProps = AccountId & TextProps;
+interface AccountAddressProps extends TextProps {
+  fingerprint: string;
+  accountIndex: number;
+  displayPreference?: AccountDisplayPreference;
+}
 
-export function AccountAddress({ accountIndex, fingerprint, ...textProps }: AccountAddressProps) {
-  const displayAddress = useAccountDisplayAddress(fingerprint, accountIndex);
+export function AccountAddress({
+  accountIndex,
+  fingerprint,
+  displayPreference,
+  ...textProps
+}: AccountAddressProps) {
+  const displayAddress = useAccountDisplayAddress({ fingerprint, accountIndex, displayPreference });
   return (
     <Text variant="label03" color="ink.text-subdued" textTransform="uppercase" {...textProps}>
       {displayAddress}
